reworked arch-dependent memory (dirty)
