+++ kernel/src/library/architectures/_x86_64/cpu/exceptions.rs
@@ -203,7 +203,7 @@ pub(super) mod handlers
 			x86_64::registers::control::Cr2::read(),
 			error_code
 		);
-		// unimplemented!();
+		unimplemented!();
 	}
 
 	/// ### CPU Exception - Security
